# AspNetCore7EFCoreDemo/Models/Course.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;
using System.Text.Json.Serialization;

namespace AspNetCore7EFCoreDemo.Models;

public partial class Course
{
    public int CourseId { get; set; }

    public string Title { get; set; }

    public int Credits { get; set; }

    public int DepartmentId { get; set; }

    [JsonIgnore]
    public virtual Department Department { get; set; }

    [JsonIgnore]
    public virtual ICollection<Enrollment> Enrollment { get; } = new List<Enrollment>();

    [JsonIgnore]
    public virtual ICollection<Person> Instructor { get; } = new List<Person>();
}